azure_functions/event-grid-trigger-function/requirements.txt
azure-functions
Pillow
requests
azure-storage-blob